% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDEsizeApp.R
\name{BDEsizeApp}
\alias{BDEsizeApp}
\title{Shiny App for efficient determination of the size of experiments in Balanced design of experiments}
\usage{
BDEsizeApp()
}
\description{
Shiny App for efficient determination of the size of experiments in Balanced design of experiments
}
\examples{
#BDEsizeApp()
}
