% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizelist.R
\name{sizelist}
\alias{sizelist}
\title{Buliding the model}
\usage{
sizelist(factor,order)
}
\arguments{
\item{factor}{the number of factor}

\item{order}{building the model with main or including the interaction effects ;
1 : only main effects(default) , 2 : both main and two-way interaction effects}
}
\value{
terms and expansion of model
}
\description{
Model is built on the number of factor and order.
}
\examples{
#2 factors; both main and two-way interaction effects
A<-sizelist(2,2)
A$full_list
A$list1

}
