% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizelist.split.R
\name{sizelist.split}
\alias{sizelist.split}
\title{Buliding the model for split-plot design}
\usage{
sizelist.split(whole.factor,split.factor,order)
}
\arguments{
\item{whole.factor}{the number of whole factor}

\item{split.factor}{the number of split factor}

\item{order}{building the model with main or including the interaction effects ;
1 : only main effects(default) , 2 : both main and two-way interaction effects}
}
\value{
terms and expansion of model
}
\description{
Model is built on the number of factor and order.
}
\examples{
#one whole.factor and one split.factor ; both main and two-way interaction effects
A<-sizelist.split(1,1,2)
A$full_list
A$list1

}
